move FileLoader and Demo to upload package
